backend/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.30.1
sqlalchemy==2.0.31
pandas==2.2.2
python-dotenv==1.0.1
pymysql==1.1.1
openpyxl==3.1.5
xlrd==2.0.1
odfpy==1.4.1

